requirements.txt
typer
httpx
rich
python-openstackclient
openstacksdk
pytest
python-dotenv